fix: mypage & login page

+++ Mallery4/app/src/main/java/com/example/mallery4/datamodel/Logindata.kt
@@ -6,5 +6,5 @@ data class Logindata (
     @SerializedName("grantType") var grantType:String,
     @SerializedName("accessToken") var accessToken:String,
     @SerializedName("refreshToken") var refreshToken:String,
-    @SerializedName("refreshTokenExpirationTime") var refreshTokenExpirationTime:Int
+    @SerializedName("refreshTokenExpirationTime") var refreshTokenExpirationTime:Long
     )
